chore: restructure package imports to enable cross package documentation
