+++ internal/analytics/doc.go
@@ -1,0 +1,4 @@
+// This package contains all the different analytics solution
+// the CLI supports
+// All those analytics can be disabled using TODO
+package analytics
