+++ internal/analytics/matomo/matomo.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/scaleway/scaleway-cli/internal/terminal"
+	"github.com/scaleway/scaleway-cli/internal/ui/terminal"
 	"github.com/scaleway/scaleway-sdk-go/scw"
 )
