+++ internal/core/bootstrap_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/alecthomas/assert"
 	"github.com/scaleway/scaleway-cli/internal/args"
-	"github.com/scaleway/scaleway-cli/internal/interactive"
+	"github.com/scaleway/scaleway-cli/internal/ui/interactive"
 )
 
 func TestInterruptError(t *testing.T) {
