+++ internal/core/cobra_usage_builder.go
@@ -10,7 +10,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/scaleway/scaleway-cli/internal/args"
-	"github.com/scaleway/scaleway-cli/internal/interactive"
+	"github.com/scaleway/scaleway-cli/internal/ui/interactive"
 	"github.com/scaleway/scaleway-sdk-go/logger"
 )
