+++ internal/namespaces/init/init.go
@@ -10,10 +10,10 @@ import (
 	"github.com/fatih/color"
 	"github.com/scaleway/scaleway-cli/internal/account"
 	"github.com/scaleway/scaleway-cli/internal/core"
-	"github.com/scaleway/scaleway-cli/internal/interactive"
 	accountcommands "github.com/scaleway/scaleway-cli/internal/namespaces/account/v2alpha1"
 	"github.com/scaleway/scaleway-cli/internal/namespaces/autocomplete"
-	"github.com/scaleway/scaleway-cli/internal/terminal"
+	"github.com/scaleway/scaleway-cli/internal/ui/interactive"
+	"github.com/scaleway/scaleway-cli/internal/ui/terminal"
 	"github.com/scaleway/scaleway-sdk-go/logger"
 	"github.com/scaleway/scaleway-sdk-go/scw"
 	"github.com/scaleway/scaleway-sdk-go/validation"
