+++ internal/namespaces/instance/v1/custom.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 
 	"github.com/scaleway/scaleway-cli/internal/core"
-	"github.com/scaleway/scaleway-cli/internal/human"
+	"github.com/scaleway/scaleway-cli/internal/ui/human"
 	"github.com/scaleway/scaleway-sdk-go/api/instance/v1"
 )
