+++ internal/namespaces/k8s/v1/custom.go
@@ -2,7 +2,7 @@ package k8s
 
 import (
 	"github.com/scaleway/scaleway-cli/internal/core"
-	"github.com/scaleway/scaleway-cli/internal/human"
+	"github.com/scaleway/scaleway-cli/internal/ui/human"
 	k8s "github.com/scaleway/scaleway-sdk-go/api/k8s/v1"
 )
