+++ internal/namespaces/registry/v1/custom_namespace.go
@@ -2,7 +2,7 @@ package registry
 
 import (
 	"github.com/fatih/color"
-	"github.com/scaleway/scaleway-cli/internal/human"
+	"github.com/scaleway/scaleway-cli/internal/ui/human"
 	"github.com/scaleway/scaleway-sdk-go/api/registry/v1"
 )
