+++ internal/ui/human/marshal_func.go
@@ -11,7 +11,7 @@ import (
 	"github.com/dustin/go-humanize"
 	"github.com/fatih/color"
 	"github.com/hashicorp/go-version"
-	"github.com/scaleway/scaleway-cli/internal/terminal"
+	"github.com/scaleway/scaleway-cli/internal/ui/terminal"
 	"github.com/scaleway/scaleway-sdk-go/scw"
 )
