+++ internal/ui/interactive/error.go
