+++ internal/ui/interactive/print.go
@@ -6,7 +6,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/scaleway/scaleway-cli/internal/terminal"
+	"github.com/scaleway/scaleway-cli/internal/ui/terminal"
 )
 
 func Print(a ...interface{}) (int, error) {
