+++ internal/ui/interactive/print_test.go
