+++ internal/ui/interactive/prompt_test.go
