+++ internal/ui/interactive/utils.go
